More cleanup & Win64 fixes
Change the FIL_PAGE_TYPE format so that bits 8..14 are reserved for
future use and must be zero.

fil_space_encrypt(): Use less stack. This seems to fix the occasional
crashes in encryption.innodb-encryption-alter.

buf_page_full_crc32_get_size(): Renamed
from buf_page_compress_fcrc32_get_size().
FIXME: Return a multiple of 256!
FIXME: Return flags: "page is compressed", "page is corrupted".

FIXME: Ensure that we always write an uncompressed page if the
compressed size is equal to innodb_page_size.

FIXME: Ensure that we are invoking the compression routines so
that a minimal-size header is created. Do not allocate 1 extra
byte for those compression algorithms that detect EOF of the
compressed stream themselves.